Cleanup get_k
